Separado da main class os requests http
-- Classes [RootHandler, AppHandler] separados em arquivos dentro da pasta http
