+++ src/http/HttpAppHandler.java
@@ -1,0 +1,4 @@
+package http;
+
+public class HttpAppHandler {
+}
